requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.1
transformers==4.45.2
torch==2.3.1
torchaudio==2.3.1
numpy
scipy
pydub==0.25.1
datasets==2.19.1
speechbrain==0.5.16
